# docs/orchestra_openapi_v2.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Orchestra API",
    "version": "1.3.0",
    "description": "API Orchestra с поддержкой SuiteRun, Planner+Resolver и Durable Execution Engine."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local dev"
    }
  ],
  "paths": {
...
    "/scenarios/{scenarioId}": {
      "get": {
        "tags": [
          "Scenarios"
        ],
        "summary": "Детали сценария",
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestScenarioDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/scenario-suites": {
      "get": {
        "tags": [
          "ScenarioSuites"
        ],
        "summary": "Список ScenarioSuite",
        "parameters": [
          {
            "name": "processId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "processVersion",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScenarioSuiteSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ScenarioSuites"
        ],
        "summary": "Создать ScenarioSuite вручную",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioSuiteCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Создан ScenarioSuite",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioSuiteDetail"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/scenario-suites/{suiteId}": {
      "get": {
        "tags": [
          "ScenarioSuites"
        ],
        "summary": "Детали ScenarioSuite",
        "parameters": [
          {
            "name": "suiteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioSuiteDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/testruns": {
      "get": {
        "tags": [
          "TestRuns"
        ],
        "summary": "Список прогонов",
        "parameters": [
          {
            "name": "scenarioId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "suiteId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "QUEUED",
                "IN_PROGRESS",
                "PASSED",
                "FAILED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "RUN_ALL_STEPS",
                "STOP_ON_FIRST_FAILURE",
                "SINGLE_STEP",
                "FROM_STEP"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestRunSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TestRuns"
        ],
        "summary": "Запуск прогона сценария",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestRunCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Прогон принят в обработку (статус PENDING/QUEUED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRunSummary"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/testruns/{testRunId}": {
      "get": {
        "tags": [
          "TestRuns"
        ],
        "summary": "Детали прогона",
        "parameters": [
          {
            "name": "testRunId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRunDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ScenarioStep": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderIndex": {
            "type": "integer"
          },
          "alias": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "description": "Тип шага: ACTION, ASSERTION или BARRIER",
            "enum": [
              "ACTION",
              "ASSERTION",
              "BARRIER"
            ]
          },
          "channelType": {
            "type": "string",
            "description": "Канал взаимодействия шага",
            "enum": [
              "HTTP_REST",
              "KAFKA",
              "GRPC",
              "DB",
              "QUEUE"
            ]
          },
          "endpointRef": {
            "$ref": "#/components/schemas/EndpointRef"
          },
          "action": {
            "type": "object",
            "properties": {
              "inputTemplate": {
                "type": "object"
              },
              "mode": {
                "type": "string",
                "enum": [
                  "SYNC",
                  "FIRE_AND_FORGET"
                ],
                "default": "SYNC",
                "description": "Режим ACTION-шагов"
              },
              "meta": {
                "description": "Мета-информация для шага (HTTP/DB/Kafka/Queue) и инструкции экспорта данных.",
                "type": "object",
                "properties": {
                  "exportAs": {
                    "type": "object",
                    "description": "Карта переменных SuiteRun.context: ключ — имя переменной, значение — JSONPath к данным ответа шага.",
                    "additionalProperties": {
                      "type": "string"
                    }
                  }
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/HttpActionMeta"
                  },
                  {
                    "$ref": "#/components/schemas/DbActionMeta"
                  },
                  {
                    "$ref": "#/components/schemas/KafkaActionMeta"
                  },
                  {
                    "$ref": "#/components/schemas/QueueActionMeta"
                  },
                  {
                    "type": "object"
                  }
                ],
                "additionalProperties": true
              }
            }
          },
          "expectations": {
            "$ref": "#/components/schemas/StepExpectations"
          }
        },
        "required": [
          "alias",
          "channelType",
          "id",
          "kind",
          "name"
        ]
      },
      "HttpActionMeta": {
        "type": "object",
        "description": "Мета-информация для HTTP-шагов.",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "HTTP"
            ]
          },
          "timeoutMs": {
            "type": "integer",
            "format": "int64"
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "additionalProperties": true
      },
      "DbActionMeta": {
        "type": "object",
        "description": "Мета-информация для шагов проверки БД.",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "DB"
            ]
          },
          "dataSource": {
            "type": "string"
          },
          "sql": {
            "type": "string"
          },
          "timeoutMs": {
            "type": "integer",
            "format": "int64"
          },
          "pollIntervalMs": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "dataSource",
          "sql"
        ],
        "additionalProperties": true
      },
      "KafkaActionMeta": {
        "type": "object",
        "description": "Мета-информация для шагов проверки Kafka.",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "KAFKA_ASSERT"
            ]
          },
          "clusterAlias": {
            "type": "string"
          },
          "topic": {
            "type": "string"
          },
          "keyExpression": {
            "type": "string"
          },
          "valueMatcherMode": {
            "type": "string",
            "enum": [
              "EXISTS_ONLY",
              "PARTIAL_MATCH"
            ]
          },
          "timeoutMs": {
            "type": "integer",
            "format": "int64"
          },
          "pollIntervalMs": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "clusterAlias",
          "topic"
        ],
        "additionalProperties": true
      },
      "QueueActionMeta": {
        "type": "object",
        "description": "Мета-информация для шагов проверки очередей.",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "QUEUE_ASSERT"
            ]
          },
          "brokerAlias": {
            "type": "string"
          },
          "queueName": {
            "type": "string"
          },
          "correlationIdExpr": {
            "type": "string"
          },
          "timeoutMs": {
            "type": "integer",
            "format": "int64"
          },
          "pollIntervalMs": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "brokerAlias",
          "queueName"
        ],
        "additionalProperties": true
      },
...
      "EndpointRef": {
        "type": "object",
        "properties": {
          "protocolId": {
            "type": "string"
          },
          "serviceName": {
            "type": "string"
          },
          "endpointName": {
            "type": "string"
          }
        },
        "required": [
          "protocolId",
          "serviceName",
          "endpointName"
        ]
      },
      "StepExpectations": {
        "type": "object",
        "properties": {
          "expectedStatusCode": {
            "type": "integer"
          },
          "businessRules": {
            "type": "object"
          },
          "schemaRefs": {
            "type": "object"
          }
        }
      },
      "TestScenarioSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "suiteId": {
            "type": "string",
            "format": "uuid",
            "description": "Идентификатор ScenarioSuite, к которому привязан сценарий"
          },
          "processId": {
            "type": "string",
            "format": "uuid"
          },
          "processVersion": {
            "type": "integer"
          },
          "key": {
            "type": "string",
            "description": "Бизнес-ключ сценария, например loan-approval:happy-path:vip"
          },
          "active": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "description": "Версия сценария"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "DEPRECATED"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "key"
        ]
      },
      "TestScenarioDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "suiteId": {
            "type": "string",
            "format": "uuid"
          },
          "key": {
            "type": "string"
          },
          "processId": {
            "type": "string",
            "format": "uuid"
          },
          "processVersion": {
            "type": "integer"
          },
          "active": {
            "type": "boolean"
          },
          "version": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "DEPRECATED"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "recommendedDataSetIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScenarioStep"
            }
          },
          "dependsOn": {
            "type": "array",
            "description": "Зависимости от других сценариев в SuiteRun.",
            "items": {
              "type": "object",
              "properties": {
                "scenarioKey": {
                  "type": "string"
                },
                "onStatus": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "PASSED",
                      "FAILED"
                    ]
                  }
                }
              },
              "required": [
                "scenarioKey",
                "onStatus"
              ]
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "key"
        ]
      },
      "ScenarioSuiteSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "processId": {
            "type": "string",
            "format": "uuid"
          },
          "processVersion": {
            "type": "integer"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "processId"
        ]
      },
      "ScenarioSuiteDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ScenarioSuiteSummary"
          },
          {
            "type": "object",
            "properties": {
              "description": {
                "type": "string"
              },
              "scenarios": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TestScenarioSummary"
                }
              }
            }
          }
        ]
      },
      "ScenarioSuiteCreateRequest": {
        "type": "object",
        "properties": {
          "processId": {
            "type": "string",
            "format": "uuid"
          },
          "processVersion": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "processId",
          "name"
        ]
      },
...
      "TestDataSetSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "scope": {
            "type": "string",
            "enum": [
              "GLOBAL",
              "SUITE",
              "SCENARIO"
            ]
          },
          "suiteId": {
            "type": "string",
            "format": "uuid"
          },
          "scenarioId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "origin": {
            "type": "string",
            "enum": [
              "MANUAL",
              "AI_GENERATED",
              "IMPORTED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "scope",
          "name",
          "origin"
        ]
      },
...
      "Violation": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Тип нарушения",
            "enum": [
              "HTTP_ERROR",
              "CONTRACT_VIOLATION",
              "BUSINESS_RULE_VIOLATION",
              "TIMEOUT",
              "VALIDATION_ERROR",
              "DB_ASSERTION_FAILED",
              "KAFKA_MESSAGE_NOT_FOUND",
              "QUEUE_MESSAGE_NOT_FOUND"
            ]
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        },
        "required": [
          "type"
        ]
      },
      "StepResult": {
        "type": "object",
        "properties": {
          "stepId": {
            "type": "string",
            "format": "uuid"
          },
          "stepAlias": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "PASSED",
              "FAILED",
              "SKIPPED",
              "FLAKY"
            ]
          },
          "durationMs": {
            "type": "integer",
            "format": "int64"
          },
          "actualStatusCode": {
            "type": "integer"
          },
          "responseBody": {
            "type": "object"
          },
          "responseHeaders": {
            "type": "object"
          },
          "dbResult": {
            "type": "object"
          },
          "kafkaMessage": {
            "type": "object"
          },
          "queueMessage": {
            "type": "object"
          },
          "violations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Violation"
            }
          }
        },
        "required": [
          "stepAlias",
          "status"
        ]
      },
      "TestRunSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "scenarioId": {
            "type": "string",
            "format": "uuid"
          },
          "scenarioVersion": {
            "type": "integer"
          },
          "suiteId": {
            "type": "string",
            "format": "uuid"
          },
          "suiteRunId": {
            "type": "string",
            "format": "uuid"
          },
          "dataSetId": {
            "type": "string",
            "format": "uuid"
          },
          "environmentId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "QUEUED",
              "IN_PROGRESS",
              "PASSED",
              "FAILED",
              "FAILED_STUCK",
              "CANCELLED"
            ]
          },
          "mode": {
            "type": "string",
            "enum": [
              "RUN_ALL_STEPS",
              "STOP_ON_FIRST_FAILURE",
              "SINGLE_STEP",
              "FROM_STEP"
            ],
            "default": "RUN_ALL_STEPS"
          },
          "startStepIndex": {
            "type": "integer",
            "minimum": 0
          },
          "targetStepAlias": {
            "type": "string"
          },
          "baseRunId": {
            "type": "string",
            "format": "uuid",
            "description": "Ссылка на прогон, из которого переиспользуется ExecutionContext"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "triggeredBy": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "scenarioId",
          "status"
        ]
      },
      "TestRunDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TestRunSummary"
          },
          {
            "type": "object",
            "properties": {
              "dataSet": {
                "$ref": "#/components/schemas/TestDataSetSummary"
              },
              "stepResults": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/StepResult"
                }
              }
            }
          }
        ]
      },
...
      "TestRunCreateRequest": {
        "type": "object",
        "properties": {
          "scenarioId": {
            "type": "string",
            "format": "uuid"
          },
          "scenarioVersion": {
            "type": "integer",
            "description": "Явная версия сценария; если не указана, используется опубликованная"
          },
          "environmentId": {
            "type": "string",
            "format": "uuid"
          },
          "dataSetId": {
            "type": "string",
            "format": "uuid"
          },
          "dataSetTag": {
            "type": "string",
            "description": "Позволяет подобрать набор данных по тегу"
          },
          "baseRunId": {
            "type": "string",
            "format": "uuid",
            "description": "Прогон, из которого нужно переиспользовать ExecutionContext"
          },
          "startStepIndex": {
            "type": "integer",
            "minimum": 0,
            "description": "Используется для режима FROM_STEP"
          },
          "targetStepAlias": {
            "type": "string",
            "description": "Используется для режима SINGLE_STEP"
          },
          "mode": {
            "type": "string",
            "enum": [
              "RUN_ALL_STEPS",
              "STOP_ON_FIRST_FAILURE",
              "SINGLE_STEP",
              "FROM_STEP"
            ],
            "default": "RUN_ALL_STEPS"
          }
        },
        "required": [
          "scenarioId"
        ]
      },
...
    }
  }
}