# services/orchestra-api/src/main/java/com/orchestra/api/dto/ScenarioSuiteDetail.java
package com.orchestra.api.dto;

import lombok.Data;
import lombok.EqualsAndHashCode;

import java.util.List;

@Data
@EqualsAndHashCode(callSuper = true)
public class ScenarioSuiteDetail extends ScenarioSuiteSummary {
    private String description;
    private List<TestScenarioSummary> scenarios;
}